fix(playground): separate permissions, tool components
