+++ packages/playground/src/components/workspace/workspace-mcp-list.tsx
@@ -1,28 +1,31 @@
 import {
 	AlertCircle,
+	ImageIcon,
 	SquareArrowOutUpRightIcon,
-	WrenchIcon,
 } from "lucide-react";
 import { useMemo } from "react";
 import { useTranslation } from "@semoss/i18n";
 import { useInsight, usePixel } from "@semoss/sdk/react";
-import type { Role } from "@semoss/shared";
 import {
 	Badge,
 	Button,
+	Card,
+	CardContent,
 	Muted,
 	ScrollArea,
-	Skeleton,
 	Tooltip,
 	TooltipContent,
 	TooltipTrigger,
 	toast,
 } from "@semoss/ui/next";
 import { mcpToPlatformUrl } from "@/components";
-import type { MCPTool } from "@/types";
 import { toSentenceCase } from "@/utility";
-
-const SHOW_MCP_LINK = import.meta.env.VITE_SHOW_MCP_LINK !== "false";
+import { AppToolSection } from "./app-tool-section";
+import {
+	type EffectivePermission,
+	PERMISSION_VARIANT,
+	type ProjectDependency,
+} from "./workspace-permissions";
 
 export interface WorkspaceMCPListProps {
 	/**
@@ -41,240 +44,9 @@ export interface WorkspaceMCPListProps {
 	search: string;
 }
 
-interface ProjectDependency {
-	engine_type:
-		| "PROJECT"
-		| "STORAGE"
-		| "DATABASE"
-		| "FUNCTION"
-		| "MODEL"
-		| "VECTOR";
-	engine_id: string;
-	engine_name: string;
-	description?: string;
-	engine_discoverable?: boolean;
-	permission_name?: Role;
-	engine_global?: boolean;
-	access_permission?: number;
-	tags: string; // comma separated tags
-}
-
-type EffectivePermission =
-	| Role
-	| "REQUESTED"
-	| "DISCOVERABLE"
-	| "FULLY_PRIVATE";
-
-const PERMISSION_VARIANT: Record<
-	EffectivePermission,
-	"default" | "secondary" | "outline" | "destructive"
-> = {
-	OWNER: "default",
-	EDIT: "secondary",
-	READ_ONLY: "outline",
-	REQUESTED: "outline",
-	DISCOVERABLE: "outline",
-	FULLY_PRIVATE: "destructive",
-};
-
-interface AppToolSectionProps {
-	dep: ProjectDependency;
-	search: string;
-	effectivePermission: EffectivePermission;
-	label: string;
-	type: "TOOLBOX" | "KNOWLEDGE";
-	onRequestAccess: (m: ProjectDependency) => void;
-}
-
-/**
- * Renders a single app's MCP tools as a collapsible section.
- * Has its own usePixel call so each app loads independently.
- */
-const AppToolSection = ({
-	dep,
-	search,
-	effectivePermission,
-	label,
-	type,
-	onRequestAccess,
-}: AppToolSectionProps) => {
-	const { t } = useTranslation("workspace");
-
-	const hasAccess =
-		effectivePermission === "READ_ONLY" ||
-		effectivePermission === "EDIT" ||
-		effectivePermission === "OWNER";
-
-	const getTools = usePixel<{ tools: MCPTool[] }>(
-		hasAccess ? `GetMCPTools("${dep.engine_id}");` : "",
-		{ data: { tools: [] } },
-	);
-
-	const appNameMatches = useMemo(
-		() =>
-			!search ||
-			dep.engine_name.toLowerCase().includes(search.toLowerCase()),
-		[dep.engine_name, search],
-	);
-
-	const filteredTools = useMemo(() => {
-		const tools = getTools.data?.tools || [];
-		// If no search, or if the app name itself matches, show all tools
-		if (!search || appNameMatches) return tools;
-		return tools.filter(
-			(tool) =>
-				(tool.title || tool.name)
-					.toLowerCase()
-					.includes(search.toLowerCase()) ||
-				(tool.description || "")
-					.toLowerCase()
-					.includes(search.toLowerCase()),
-		);
-	}, [getTools.data, search, appNameMatches]);
-
-	// Hide section if searching and nothing matches (only after tools have loaded)
-	if (
-		hasAccess &&
-		search &&
-		!appNameMatches &&
-		getTools.status === "SUCCESS" &&
-		filteredTools.length === 0
-	) {
-		return null;
-	}
-
-	const accessMissing =
-		effectivePermission === "REQUESTED" ||
-		effectivePermission === "DISCOVERABLE" ||
-		effectivePermission === "FULLY_PRIVATE";
-
-	return (
-		<div
-			className={`rounded-lg border ${
-				accessMissing
-					? "border-destructive/50 border-dashed"
-					: "border-border"
-			}`}
-		>
-			{/* Section header */}
-			<div className="flex items-center gap-2 border-b px-4 py-2">
-				<div className="min-w-0 flex-1">
-					<span className="truncate font-semibold text-sm">
-						{dep.engine_name}
-					</span>
-				</div>
-
-				<Badge variant="outline" className="shrink-0 text-xs">
-					{toSentenceCase(dep.engine_type)}
-				</Badge>
-
-				{effectivePermission === "DISCOVERABLE" ? (
-					<Button
-						size="sm"
-						className="h-fit shrink-0 px-2 py-1 text-xs"
-						onClick={() => onRequestAccess(dep)}
-					>
-						{label}
-					</Button>
-				) : (
-					<Badge
-						variant={PERMISSION_VARIANT[effectivePermission]}
-						className="shrink-0 text-xs"
-					>
-						{label}
-					</Badge>
-				)}
-
-				{SHOW_MCP_LINK &&
-					(effectivePermission === "FULLY_PRIVATE" ? (
-						<Tooltip>
-							<TooltipTrigger asChild>
-								<AlertCircle className="size-4 shrink-0 cursor-help text-destructive" />
-							</TooltipTrigger>
-							<TooltipContent>
-								{t("mcp.tooltipNoAccess", {
-									type:
-										type === "TOOLBOX"
-											? "toolbox"
-											: "knowledge base",
-								})}
-							</TooltipContent>
-						</Tooltip>
-					) : (
-						<Tooltip>
-							<TooltipTrigger asChild>
-								<Button
-									variant="ghost"
-									size="icon"
-									className={`-mr-2 size-8 shrink-0 ${
-										accessMissing ? "text-destructive" : ""
-									}`}
-									asChild
-								>
-									<a
-										target="_blank"
-										href={mcpToPlatformUrl(dep)}
-									>
-										<SquareArrowOutUpRightIcon className="size-4" />
-									</a>
-								</Button>
-							</TooltipTrigger>
-							<TooltipContent>
-								{t("mcp.tooltipOpen", {
-									type:
-										type === "TOOLBOX"
-											? "toolbox"
-											: "knowledge base",
-								})}
-							</TooltipContent>
-						</Tooltip>
-					))}
-			</div>
-
-			{/* Tools body */}
-			{!hasAccess ? (
-				<p className="px-4 py-3 text-muted-foreground text-sm italic">
-					{label}
-				</p>
-			) : getTools.status === "LOADING" ||
-				getTools.status === "INITIAL" ? (
-				<div className="space-y-3 p-4">
-					<Skeleton className="h-4 w-3/4" />
-					<Skeleton className="h-4 w-1/2" />
-					<Skeleton className="h-4 w-2/3" />
-				</div>
-			) : filteredTools.length > 0 ? (
-				<ul className="divide-y divide-border">
-					{filteredTools.map((tool) => (
-						<li
-							key={tool.name}
-							className="flex items-start gap-3 px-4 py-3"
-						>
-							<WrenchIcon className="mt-0.5 size-4 shrink-0 text-muted-foreground" />
-							<div className="min-w-0 flex-1">
-								<div className="font-medium text-sm leading-snug">
-									{tool.title || tool.name}
-								</div>
-								{tool.description && (
-									<div className="mt-0.5 text-muted-foreground text-xs">
-										{tool.description}
-									</div>
-								)}
-							</div>
-						</li>
-					))}
-				</ul>
-			) : (
-				<p className="px-4 py-3 text-muted-foreground text-sm italic">
-					{t("mcp.noDescription")}
-				</p>
-			)}
-		</div>
-	);
-};
-
 /**
  * Renders the list of MCP apps and their individual tools for a workspace.
+ * Knowledge tab uses a card grid; Toolbox tab uses a tool drill-down list.
  *
  * @component
  */
@@ -301,16 +73,23 @@ export const WorkspaceMCPList = ({
 		},
 	);
 
-	// Filter by type only — search is handled per-section in AppToolSection
 	const filteredDeps = useMemo(() => {
-		return (
+		const byType =
 			getDependencies.data?.filter((m) =>
 				type === "TOOLBOX"
 					? m.engine_type !== "VECTOR"
 					: m.engine_type === "VECTOR",
-			) || []
-		);
-	}, [getDependencies.data, type]);
+			) || [];
+		// Knowledge tab filters search inline; Toolbox delegates to AppToolSection
+		if (type === "KNOWLEDGE" && search) {
+			return byType.filter(
+				(m) =>
+					m.engine_id.toLowerCase().includes(search.toLowerCase()) ||
+					m.engine_name.toLowerCase().includes(search.toLowerCase()),
+			);
+		}
+		return byType;
+	}, [getDependencies.data, type, search]);
 
 	const getEffectivePermission = (
 		m: ProjectDependency,
@@ -381,6 +160,150 @@ export const WorkspaceMCPList = ({
 						: t("mcp.noKnowledge")}
 				</Muted>
 			</div>
+		);
+	}
+
+	if (type === "KNOWLEDGE") {
+		return (
+			<ScrollArea className="h-full w-full">
+				<div className="grid grid-cols-1 gap-4 p-4 md:grid-cols-2 lg:grid-cols-3">
+					{filteredDeps.map((m) => {
+						const { effectivePermission, label } =
+							getEffectivePermission(m);
+						const accessMissing =
+							effectivePermission === "REQUESTED" ||
+							effectivePermission === "DISCOVERABLE" ||
+							effectivePermission === "FULLY_PRIVATE";
+						return (
+							<Card
+								key={m.engine_id}
+								className={`col-span-1 p-0 ${
+									accessMissing
+										? "border-destructive/50 border-dashed"
+										: ""
+								}`}
+							>
+								<CardContent className="space-y-2 p-4">
+									{/* Title & Open Button */}
+									<div className="flex items-start justify-between gap-2">
+										<div className="wrap-break-word min-w-0 flex-1 font-semibold text-sm leading-tight">
+											{m.engine_name}
+										</div>
+										<Tooltip>
+											<TooltipTrigger asChild>
+												{effectivePermission ===
+												"FULLY_PRIVATE" ? (
+													<AlertCircle className="size-4 shrink-0 cursor-help text-destructive" />
+												) : (
+													<Button
+														variant="ghost"
+														size="icon"
+														className={`-m-2 shrink-0 ${
+															accessMissing
+																? "text-destructive"
+																: ""
+														}`}
+														asChild
+													>
+														<a
+															target="_blank"
+															href={mcpToPlatformUrl(
+																m,
+															)}
+														>
+															<SquareArrowOutUpRightIcon className="size-4" />
+														</a>
+													</Button>
+												)}
+											</TooltipTrigger>
+											<TooltipContent>
+												{accessMissing
+													? t("mcp.tooltipNoAccess", {
+															type: "knowledge base",
+														})
+													: t("mcp.tooltipOpen", {
+															type: "knowledge base",
+														})}
+											</TooltipContent>
+										</Tooltip>
+									</div>
+
+									{/* Image & Details */}
+									<div className="flex items-center gap-3">
+										{effectivePermission ===
+										"FULLY_PRIVATE" ? (
+											<div className="flex size-16 shrink-0 items-center justify-center rounded-md border border-border border-dashed bg-muted/50">
+												<ImageIcon className="size-6 text-muted-foreground" />
+											</div>
+										) : (
+											<img
+												src={`${import.meta.env.MODULE}/api/e-${m.engine_id}/image/download`}
+												alt={m.engine_name}
+												className="size-16 shrink-0 rounded-md object-cover object-center"
+											/>
+										)}
+
+										<div className="flex flex-1 flex-col gap-2">
+											<Badge
+												variant="outline"
+												className="w-fit"
+											>
+												{toSentenceCase(m.engine_type)}
+											</Badge>
+
+											{effectivePermission ===
+											"DISCOVERABLE" ? (
+												<Button
+													size="sm"
+													className="h-fit w-fit px-2 py-1 text-xs"
+													onClick={() =>
+														handleRequestAccess(m)
+													}
+												>
+													{t(
+														"mcp.requestAccessButton",
+													)}
+												</Button>
+											) : (
+												<Badge
+													variant={
+														PERMISSION_VARIANT[
+															effectivePermission
+														]
+													}
+													className="w-fit"
+												>
+													{label}
+												</Badge>
+											)}
+										</div>
+									</div>
+
+									{/* Description */}
+									<div className="text-muted-foreground text-xs">
+										{m.description ||
+											t("mcp.noDescription")}
+									</div>
+
+									{m.tags?.length > 0 && (
+										<div className="flex flex-wrap gap-1">
+											{m.tags.split(",").map((tag) => (
+												<Badge
+													key={tag}
+													variant="secondary"
+													className="text-xs"
+												>
+													{tag}
+												</Badge>
+											))}
+										</div>
+									)}
+								</CardContent>
+							</Card>
+						);
+					})}
+				</div>
+			</ScrollArea>
 		);
 	}
